requirements.txt
anyio==3.7.1
blinker==1.6.2
certifi==2023.7.22
click==8.1.6
django-environ==0.10.0
dnspython==2.4.0
exceptiongroup==1.1.2
Flask==2.3.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.4.1
sniffio==1.3.0
Werkzeug==2.3.6
zipp==3.16.2
